Fixed a bug regarding nullability annotations, when running with the latest preview version of `Microsoft.CodeAnalysis.NetAnalyzers`.
Basically, putting a null-check-override on something once is no longer sufficient to eliminate null-checks down the line, when extensions are used. In truth, a null-check-override was never intended to carry down the line to future uses, but it appeared this way because Roslyn would incorrectly mark values that have the . operator applied to them as not possibly null, not realizing that extension methods make it possible to invoke the . operator upon null.

https://github.com/dotnet/roslyn/issues/48605