Fix send/receive gateway tasks never completing
If RunAsync was called with an already cancelled token, this delay would throw an improperly handled TaskCanceledException. As a result, regular cleanup would be bypassed.

Also fixed certain transport service disconnect calls indicating they will reconnect when they won't.

This ensures the path of disconnection is followed as if the gateway itself requested it.