perf: Initialize reasonable starting buffer size
This commit builds on top of the previous commit by
assigning an initial buffer size when allocating the buffer writer.

This is done to avoid unnecessary resizing of the array,while also
minimizing the actual size of the buffer in the most common case.

128 was a value picked because it is a generally suitable value for
most payloads sent over the gateway, namely `HEARTBEAT`s and `IDENTIFY`s