fix: Use `WrittenMemory`, not `GetMemory()`
The method returns a memory pointing at the first free byte, which can
have garbled data depending on how the writer was last used.

Furthermore, it is incorrect to use that here, as we need to send the
portion of the buffer that has been written to.

This commit fixes that issue.