feat(commands): Add support for user commands
N.B: It is important to note that Remora allows:
- Automatic framework responses
- Those responses to be made ephemeral

Because of this, it brings into question whether there should be
validation on the account that a command may be marked as ephemeral,
(or worse, may be marked as NON-ephemeral) when the callback hint would
indicate that the opposite is true.

This behavior- or lack thereof- is technically spawned from developer
error, despite being the framework's decision on how to handle commands.