fix: Remove erroneous converter & add missing json
This removes the string enum converter for ApplicationIntegrationType,
which caused deserialization to fail due to the fact that Discord
accepted a stringified value, but would return a numeric value in the
response payload, which the enum converter cannot handle.

Missing JSON fields (`contexts`, `integration_types` have also been
added, as they were previously missing, and likely would've caught
this issue.

A new converter type was also added to fix this confounding issue, but
it breaks on `Optional<IReadOnlyDictionary<...>>` which confuses me.

This is being posted purely for debugging purposes, and is known
to be broken at this point for reasons entirely beyond me and perhaps
even God.