fix(core)!: Make IMessageUpdate inherit IMessage
This commit changes the IMessageUpdate interface (and subsequently, its
implementation)to implement the IMessageInterface instead of the
IPartialMessage interface.

This change is due to Discord returning a full message object on
MESSAGE_UPDATE.

BREAKING-CHANGE: IMessageUpdate now implements IMessage
(which implements IPartialMessage),
`Mentions` is no longer optional. The ad-hoc implementation for
IPartialMessage was kept out of convenience (and to avoid further breaking changes)