Parse channel and message urls with SnowflakeParser
If user sends a link to a channel, it's not being parsed by Remora. On
the other hand the client understands this link and visually there is
no difference between a mention and a channel link.

Add parsing of a channel links, without checking the whole string,
only beginning 'https://discord.com/channels/' is used to check this
is a url, and only the last part of the string is then used, all other
parts are ignored.